+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+**11 Mar 2022**
+
+More user-friendly UI.
+
+Fixed presale false :
+- countdown enabled for WL when presale false
+- if discountPrice is null, then mint becomes private to WL.
+
 **03 Mar 2022**
 
 endSettings now properly managed on front end :